+++ src/main/java/fr/utbm/to52/smarthome/controller/Controller.java
@@ -2,23 +2,7 @@ package fr.utbm.to52.smarthome.controller;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
-import org.json.JSONArray;
-
-import fr.utbm.to52.smarthome.controller.events.AbortEvent;
-import fr.utbm.to52.smarthome.controller.events.AddNoteEvent;
-import fr.utbm.to52.smarthome.controller.events.AddRingEvent;
-import fr.utbm.to52.smarthome.controller.events.GetLogBookEvent;
-import fr.utbm.to52.smarthome.controller.events.GetNoteEvent;
-import fr.utbm.to52.smarthome.controller.events.LightEvent;
-import fr.utbm.to52.smarthome.controller.events.MotionSensor;
-import fr.utbm.to52.smarthome.controller.events.NoSuchCommand;
-import fr.utbm.to52.smarthome.controller.events.PresenceEvent;
-import fr.utbm.to52.smarthome.controller.events.QuitEvent;
-import fr.utbm.to52.smarthome.controller.events.RingEvent;
-import fr.utbm.to52.smarthome.controller.events.ShutterEvent;
-import fr.utbm.to52.smarthome.controller.events.core.Event;
 import fr.utbm.to52.smarthome.services.AbstractService;
 import fr.utbm.to52.smarthome.services.Service;
 import fr.utbm.to52.smarthome.services.clock.ClockService;
@@ -71,14 +55,16 @@ public class Controller extends AbstractService{
 		this.MQTT = new MQTTService(this.cmdHandler);
 		this.lService.add(this.MQTT);
 		
-		this.server = new CmdServer(this.cmdHandler);
-		this.lService.add(this.server);
-		
-		this.clock = new ClockService();
-		this.lService.add(this.clock);
-		
 		this.couch = new CouchdbService();
 		this.lService.add(this.couch);
+		
+		if(this.config.getFeaturesEnabled().contains("clock")){
+			this.server = new CmdServer(this.cmdHandler);
+			this.lService.add(this.server);
+			
+			this.clock = new ClockService();
+			this.lService.add(this.clock);
+		}
 		
 		/*this.mail = new GmailService();
 		this.lService.add(this.mail);*/
@@ -97,28 +83,8 @@ public class Controller extends AbstractService{
 		this.running = true;
 	}
 	
-	private EventController eventInit;
-	
 	private void enableEvent(){ // TODO make DAO singleton
-		this.eventInit = new EventController(this);
-		for(Map.Entry<String, JSONArray> entry : this.config.getFeaturesMap().entrySet())
-			this.eventInit.initialize(entry.getValue());
 		
-		this.cmdHandler.setQuitEvent(new QuitEvent(this, this.couch.getSession()));
-		this.cmdHandler.setNoSuchCommand(new NoSuchCommand(this.couch.getSession()));
-		this.cmdHandler.setRingEventController(new RingEvent(this.couch.getSession(), this.MQTT.getMqtt()));
-		this.cmdHandler.setAddRingEventController(new AddRingEvent(this.couch.getSession(), this.clock.getCron()));
-		this.cmdHandler.setLightEvent(new LightEvent(this.couch.getSession(), this.MQTT.getMqtt()));
-		this.cmdHandler.setAddNote(new AddNoteEvent(this.couch.getSession(), this.couch.getNoteDao()));
-		this.cmdHandler.setGetNote(new GetNoteEvent(this.couch.getSession(), this.couch.getNoteDao(), this.MQTT.getMqtt()));
-		this.cmdHandler.setGetLogBook(new GetLogBookEvent(this.couch.getSession(), this.couch.getLogbookDAO(), this.MQTT.getMqtt()));
-		
-		List<Event> ev = new ArrayList<>();
-		ev.add(new AbortEvent(this.couch.getSession(), this.MQTT.getMqtt()));
-		ev.add(new ShutterEvent(this.couch.getSession(), this.MQTT.getMqtt()));
-		this.cmdHandler.setMotionSensor(new MotionSensor(this.couch.getSession(), ev));
-		
-		this.cmdHandler.setPresence(new PresenceEvent(this.couch.getSession(), this.MQTT.getMqtt()));
 	}
 	
 	@Override
