+++ src/main/java/fr/utbm/to52/smarthome/controller/events/core/AbstractDAOComEvent.java
@@ -6,7 +6,6 @@ package fr.utbm.to52.smarthome.controller.events.core;
 import org.json.JSONObject;
 import org.lightcouch.CouchDbClient;
 
-import fr.utbm.to52.smarthome.controller.Conf;
 import fr.utbm.to52.smarthome.repository.DAO;
 import fr.utbm.to52.smarthome.services.com.MQTT;
 
@@ -35,7 +34,7 @@ public abstract class AbstractDAOComEvent<N> extends AbstractDAOEvent<N> {
 	 * @param data JSONobject to answer
 	 */
 	protected void publish(String topic, JSONObject data){
-		this.mqtt.publish("/" + topic + "/" + Conf.getInstance().getTopicAnswerSuffix(), data.toString());
+		this.mqtt.publish("/" + topic, data.toString());
 	}
 	
 	/**
@@ -44,7 +43,7 @@ public abstract class AbstractDAOComEvent<N> extends AbstractDAOEvent<N> {
 	 * @param data String to send
 	 */
 	protected void publish(String topic, String data){
-		this.mqtt.publish("/" + topic + "/" + Conf.getInstance().getTopicAnswerSuffix(), data);
+		this.mqtt.publish("/" + topic, data);
 	}
 	
 	/**
